% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adj_object.R
\name{adj_object}
\alias{adj_object}
\title{Loglikelihood adjustment of fitted model objects}
\usage{
adj_object(x, cluster = NULL, use_sandwich = TRUE, use_vcov = TRUE,
  ...)
}
\arguments{
\item{x}{A fitted model object.}

\item{cluster}{see details in \code{\link[chandwich]{adjust_loglik}}.}

\item{use_sandwich}{A logical scalar. If \code{use_sandwich = TRUE}, we use
\code{\link[sandwich]{meat}} or \code{\link[sandwich]{meatCL}} to estimate the
\code{V} in \code{\link[chandwich]{adjust_loglik}}. Otherwise, set \code{V} as
\code{NULL}.}

\item{use_vcov}{A logical scalar. If \code{use_vcov = TRUE} and the method for
\code{x} exists, we use it to estimate the Hessian of the independence
loglikelihood and pass it as \code{H} to \code{\link[chandwich]{adjust_loglik}}.
Otherwise, set \code{H} as \code{NULL}.}

\item{...}{Further arguments to be passed to the functions in the
sandwich package \code{\link[sandwich]{meat}}, if \code{cluster = NULL},
or \code{\link[sandwich:vcovCL]{meatCL}}, otherwise.}
}
\value{
An object of class \code{c("oolax", "chandwich")} with the same structure as an
        object returned from \code{\link[chandwich]{adjust_loglik}}.
}
\description{
Loglikelihood adjustment of fitted objects
}
\details{
Performs loglikelihood adjustment of fitted objects for clustered data, based
on Chandler and Bata (2007). We use vertical adjustment which is described in Section 6.
}
