% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evd_fgev.R
\name{logLikVec.gev}
\alias{logLikVec.gev}
\title{logLikVec - loglikelihood vector for GEV Distribution}
\usage{
\method{logLikVec}{gev}(object, pars = NULL, ...)
}
\arguments{
\item{object}{Object of class "gev".}

\item{...}{Additional optional arguments. At present no optional arguments are used.}

\item{contrib}{If \code{contrib = TRUE} then return all conributions
to the log-likelihood  Otherwise return the total log-likelihood.}
}
\value{
An object of class \code{c("oolax", "chandwich")} with the same structure as an
        object returned from \code{\link[chandwich]{adjust_loglik}}.
}
\description{
It provides a vector of the contributions to the independence loglikelihood from
individual observations for Generalized Extreme Value distribution. Used by
\code{alogLik.evd()}.
}
\examples{
library(evd)
uvdata <- rgev(100, loc = 0.13, scale = 1.1, shape = 0.2)
M1 <- fgev(uvdata, nsloc = (-49:50)/100)
logLik(M1)
logLikVec(M1, contrib = FALSE)
}
