% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ismev_gev_fit.R
\name{logLikVec.ismev_gev}
\alias{logLikVec.ismev_gev}
\title{Adjusted loglikelihood of the GEV distribution based on \code{ismev::gev.fit}.}
\usage{
\method{logLikVec}{ismev_gev}(object, pars = NULL, ...)
}
\arguments{
\item{...}{Further arguments to be passed to the functions in the
sandwich package \code{\link[sandwich]{meat}}, if \code{cluster = NULL},
or \code{\link[sandwich:vcovCL]{meatCL}}, otherwise.}
}
\value{
An object of class \code{c("oolax", "chandwich")} with the same structure as an
        object returned from \code{\link[chandwich]{adjust_loglik}}.
}
\description{
Adjusted loglikelihood of the GEV distribution based on \code{ismev::gev.fit}.
}
