% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ismev_gpd_fit.R
\name{logLikVec.ismev_gpd}
\alias{logLikVec.ismev_gpd}
\title{Additional optional arguments. At present no optional arguments are used.
let's check github}
\usage{
\method{logLikVec}{ismev_gpd}(object, pars = NULL, ...)
}
\description{
Additional optional arguments. At present no optional arguments are used.
let's check github
}
