% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evd_fpot.R
\name{logLikVec.pot}
\alias{logLikVec.pot}
\title{logLikVec - loglikelihood vector for GPD}
\usage{
\method{logLikVec}{pot}(object, pars = NULL, ...)
}
\arguments{
\item{object}{Object of class "pot"}

\item{...}{Additional optional arguments. At present no optional arguments are used.}

\item{contrib}{If \code{contrib = TRUE} then return all conributions
to the log-likelihood  Otherwise return the total log-likelihood}
}
\value{
An object of class \code{c("oolax", "chandwich")} with the same structure as an
        object returned from \code{\link[chandwich]{adjust_loglik}}.
}
\description{
It provides a vector of the contributions to the independence loglikelihood from
individual observations for Peaks Over Threshold modelling. Used by
\code{alogLik.evd()}
}
\details{
Performs loglikelihood adjustment of fitted objects for clustered data, based
on Chandler and Bata (2007). We use vertical adjustment which is described in Section 6.
}
\examples{
library(evd)
uvdata <- rgpd(100, loc = 0, scale = 1.1, shape = 0.2)
M1 <- fpot(uvdata, 1)
logLik(M1)
logLikVec(M1, contrib = FALSE)
}
