% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oolax.R
\docType{package}
\name{oolax}
\alias{oolax}
\alias{oolax-package}
\title{oolax: Objected-orientated Loglikelihood Adjustment for Extreme Value Models}
\description{
Performs adjustments of the independence loglikelihood for fitted model
objects produced by several R packages for extreme value modelling.
The packages are \code{evd}, \code{ismev}, ...
}
\details{
The main function in the oolax package is \code{\link{alogLik}}, which provides
loglikelihood adjustment for fitting Generalised Extreme Value (GEV) model and
Threshold Modelling using generalised Pareto distribution (GPD). The adjustment is
based on package \code{\link{evd}} and \code{\link{ismev}}. The user can make
adjustment towards fitted obejcts from \code{\link{evd}} and \code{\link{evd}}. The
metholdology is develped based on  \code{\link{sandwich}} and  \code{\link{chandwich}}.

See \code{vignette("oolax-vignette", package = "oolax")} for an overview and detailed
examples about the package.
}
\references{
Chandler, R. E. and Bate, S. (2007). Inference for clustered
  data using the independence loglikelihood. \emph{Biometrika},
  \strong{94}(1), 167-183. \url{http://dx.doi.org/10.1093/biomet/asm015}
}
