% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ismev.R
\name{ismev}
\alias{ismev}
\alias{alogLik.gev.fit}
\alias{alogLik.gpd.fit}
\title{Loglikelihood adjustment of ismev fits}
\usage{
\method{alogLik}{gev.fit}(x, cluster = NULL, use_vcov = TRUE, ...)

\method{alogLik}{gpd.fit}(x, cluster = NULL, use_vcov = TRUE, ...)
}
\arguments{
\item{x}{A fitted model object.}

\item{cluster}{see details in \code{\link[chandwich]{adjust_loglik}}.}

\item{use_vcov}{A logical scalar. If \code{use_vcov = TRUE} and the method for
\code{x} exists, we use it to estimate the Hessian of the independence
loglikelihood and pass it as \code{H} to \code{\link[chandwich]{adjust_loglik}}.
Otherwise, set \code{H} as \code{NULL}.}

\item{...}{Further arguments to be passed to the functions in the
sandwich package \code{\link[sandwich]{meat}}, if \code{cluster = NULL},
or \code{\link[sandwich:vcovCL]{meatCL}}, otherwise.}
}
\value{
An object of class \code{c("oolax", "chandwich")} with the same structure as an
        object returned from \code{\link[chandwich]{adjust_loglik}}.
}
\description{
Loglikelihood adjustment for fitting Generalised Extreme Value (gev) model and
Threshold Modelling using generalised Pareto distribution (gpd). The adjustment is
based on package \code{ismev}.
}
\examples{
# We need the ismev package
got_ismev <- requireNamespace("ismev", quietly = TRUE)

if (got_ismev) {
  library(ismev)

  # An example from Chandler and Bate (2007)
  y <- c(chandwich::owtemps[, "Oxford"], chandwich::owtemps[, "Worthing"])
  x <- as.matrix(rep(c(1, -1), each = length(y) / 2))
  owfit <- oogev.fit(y, x, mul = 1, sigl = 1, shl = 1, method = "BFGS" )
  year <- rep(1:(length(y) / 2), 2)
  adj_owfit <- alogLik(owfit, cluster = year, cadjust = FALSE)
  summary(adj_owfit)

 # An example from the gpd.fit() documentation
 data(rain)
 fit <- oogpd.fit(rain, 10)
 adj_fit <- alogLik(fit)
 summary(adj_fit)
}
}
