% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evd_fgev.R
\name{logLikVec.gev}
\alias{logLikVec.gev}
\title{logLikVec - loglikelihood vector for GEV Distribution}
\usage{
\method{logLikVec}{gev}(object, pars = NULL, ...)
}
\arguments{
\item{object}{Object of class "gev".}

\item{...}{Additional optional arguments. At present no optional arguments are used.}

\item{contrib}{If \code{contrib = TRUE} then return all conributions
to the log-likelihood  Otherwise return the total log-likelihood.}
}
\value{
An object of class \code{c("oolax", "chandwich")} with the same structure as an
        object returned from \code{\link[chandwich]{adjust_loglik}}.
}
\description{
It provides a vector of the contributions to the independence loglikelihood from
individual observations for Generalized Extreme Value distribution. Used by
\code{alogLik.evd()}.
}
\details{
Performs loglikelihood adjustment of fitted objects for clustered data, based
on Chandler and Bata (2007). We use vertical adjustment which is described in Section 6
in Chandler and Bata (2007). The fitted object x must have S3 methods:
\code{logLikVec}, \code{coef}, and \code{nobs}.  It mayhave method \code{vcov} and
\code{estfun}. If a \code{vcov} method is not available then the variance-covariance
matrix of the model parameters is estimated inside
\code{\link[chandwich]{adjust_loglik}}. If an \code{estfun} method is not available
then the score matrix is estimated using \code{\link[numDeriv]{jacobian}}. More
details can be found on \code{\link[oolax]{logLikVec.gev}}.
}
\examples{
library(evd)
uvdata <- rgev(100, loc = 0.13, scale = 1.1, shape = 0.2)
M1 <- fgev(uvdata, nsloc = (-49:50)/100)
logLik(M1)
logLikVec(M1, contrib = FALSE)
}
\seealso{
\code{\link[chandwich]{adjust_loglik}} Loglikelihood adjustment using the sandwich
estimator.

\code{\link[oolax]{logLikVec.gev}} Loglikelihood vector for GEV Distribution fitted
by \code{\link[evd]{fgev}}

\code{\link[oolax]{logLikVec.pot}} Loglikelihood vector for GPD Distribution fitted
by \code{\link[evd]{fpot}}

\code{\link[oolax]{logLikVec.ismev_gev}} Loglikelihood vector for GEV Distribution
fitted by \code{\link[ismev]{gev.fit}}

\code{\link[oolax]{logLikVec.ismev_gpd}} Loglikelihood vector for GPD Distribution
fitted by \code{\link[ismev]{gpd.fit}}
}
