% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ismev_gev_fit.R
\name{logLikVec.ismev_gev}
\alias{logLikVec.ismev_gev}
\title{logLikVec - Adjusted loglikelihood vector for GEV Distribution fitted by
ismev::gev.fit}
\usage{
\method{logLikVec}{ismev_gev}(object, pars = NULL, ...)
}
\arguments{
\item{object}{Object of class "ismev_gev". It must come from
\code{\link[ismev]{gev.fit}}.}

\item{pars}{parameters for the fitted object. Since the object is of class "ismev_gev",
it is not necessary to input the parameters.}

\item{...}{Additional optional arguments. At present no optional arguments are used.}
}
\value{
An object of class \code{c("oolax", "chandwich")} with the same structure as an
        object returned from \code{\link[chandwich]{adjust_loglik}}.
}
\description{
It provides a vector of the contributions to the independence loglikelihood from
individual observations for Generalized Extreme Value distribution. Used by
\code{logLikVec()}.
}
\details{
The fitted object must have S3 methods:
\code{logLikVec}, \code{coef}, and \code{nobs}.  It may have method \code{vcov} and
\code{estfun}. If a \code{vcov} method is not available then the variance-covariance
matrix of the model parameters is estimated inside
\code{\link[chandwich]{adjust_loglik}}. If an \code{estfun} method is not available
then the score matrix is estimated using \code{\link[numDeriv]{jacobian}}. More in
See Also.
}
\seealso{
\code{\link[chandwich]{adjust_loglik}} Loglikelihood adjustment using the sandwich
estimator.

\code{\link[oolax]{logLikVec.gev}} Loglikelihood vector for GEV Distribution fitted
by \code{\link[evd]{fgev}}

\code{\link[oolax]{logLikVec.pot}} Loglikelihood vector for GPD Distribution fitted
by \code{\link[evd]{fpot}}

\code{\link[oolax]{logLikVec.ismev_gev}} Loglikelihood vector for GEV Distribution
fitted by \code{\link[ismev]{gev.fit}}

\code{\link[oolax]{logLikVec.ismev_gpd}} Loglikelihood vector for GPD Distribution
fitted by \code{\link[ismev]{gpd.fit}}
}
