% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evd_fpot.R
\name{logLikVec.pot}
\alias{logLikVec.pot}
\title{logLikVec - loglikelihood vector for GPD}
\usage{
\method{logLikVec}{pot}(object, pars = NULL, ...)
}
\arguments{
\item{object}{Object of class "pot"}

\item{...}{Additional optional arguments. At present no optional arguments are used.}

\item{contrib}{If \code{contrib = TRUE} then return all conributions
to the log-likelihood  Otherwise return the total log-likelihood}
}
\value{
An object of class \code{c("oolax", "chandwich")} with the same structure as an
        object returned from \code{\link[chandwich]{adjust_loglik}}.
}
\description{
It provides a vector of the contributions to the independence loglikelihood from
individual observations for Peaks Over Threshold modelling. Used by
\code{alogLik.evd()}
}
\details{
The fitted object x must have S3 methods:
\code{logLikVec}, \code{coef}, and \code{nobs}.  It mayhave method \code{vcov} and
\code{estfun}. If a \code{vcov} method is not available then the variance-covariance
matrix of the model parameters is estimated inside
\code{\link[chandwich]{adjust_loglik}}. If an \code{estfun} method is not available
then the score matrix is estimated using \code{\link[numDeriv]{jacobian}}. More
details can be found on \code{\link[oolax]{logLikVec.gev}}.
}
\examples{
library(evd)
uvdata <- rgpd(100, loc = 0, scale = 1.1, shape = 0.2)
M1 <- fpot(uvdata, 1)
logLik(M1)
logLikVec(M1, contrib = FALSE)
}
\seealso{
\code{\link[chandwich]{adjust_loglik}} Loglikelihood adjustment using the sandwich
estimator.

\code{\link[oolax]{logLikVec.gev}} Loglikelihood vector for GEV Distribution fitted
by \code{\link[evd]{fgev}}

\code{\link[oolax]{logLikVec.pot}} Loglikelihood vector for GPD Distribution fitted
by \code{\link[evd]{fpot}}

\code{\link[oolax]{logLikVec.ismev_gev}} Loglikelihood vector for GEV Distribution
fitted by \code{\link[ismev]{gev.fit}}

\code{\link[oolax]{logLikVec.ismev_gpd}} Loglikelihood vector for GPD Distribution
fitted by \code{\link[ismev]{gpd.fit}}
}
